Handle anchors before the start of the buffer and after the end
This commit makes anchors an enum type, so that we can introduce special
Start and End anchors. These anchors are always at the start or end of
the buffer, regardless of changes. The rest of anchors are implemented
with the "Middle" enum variant, which refers to a logical position with
a bias.